Splitting new code in separate module, updating readme and changes.txt
