+++ CHANGES.txt
@@ -3,6 +3,8 @@
     - support for eps,delta computations with varying parameters (docs, tests, code quality)
     - added experimental computations for discrete mechanisms (in `fourier_accountant.experimental`)
     - added experimental computations for tight upper/lower for subsampled Gaussian mechanism (in `fourier_accountant.experimental`)
+    - introduced PLD class structure for subsampled Gaussian and discrete mechanisms
+    - introduced routines to compute bounds on privacy parameters given a PLD instance
 
 # 0.12.0
     - first public release as `fourier_accountant` package
