+++ fourier_accountant/__init__.py
@@ -12,12 +12,7 @@ The code is due to Antti Koskela (@koskeant) and Joonas Jälkö (@jjalko) and
 was refactored by Lukas Prediger (@lumip) .
 '''
 
-from .compute_eps import get_epsilon_R, get_epsilon_S
-from .compute_delta import get_delta_R, get_delta_S
+from fourier_accountant.accountant import *
 
-__all__ = [
-    "get_epsilon_R",
-    "get_epsilon_S",
-    "get_delta_R",
-    "get_delta_S"
-]
+from fourier_accountant.compute_eps import get_epsilon_R, get_epsilon_S
+from fourier_accountant.compute_delta import get_delta_R, get_delta_S
